+++ cosim/simDiag/simPio.h
@@ -41,7 +41,7 @@ extern "C" {
   void sim_Write32_64(u_int32_t adr, u_int64_t data);
   void sim_Write64_64(u_int64_t adr, u_int64_t data);  
   void sim_WriteDvtFlag(int msb, int lsb, int val) ;
-  int  sim_ReadDvtFlag(int msb, int lsb) ;  
+  u_int64_t  sim_ReadDvtFlag(int msb, int lsb) ;  
   void sim_Framer_RdWr(u_int32_t adr, u_int32_t wrDat, u_int32_t *rdDat);
   void sim_Sample_RdWr(u_int32_t adr, u_int64_t wrDat, u_int64_t *rdDat);
   int  sim_SetInActiveStatus(void);
